Type safety

+++ src/app/api/upload/route.ts
@@ -5,29 +5,31 @@ import os from 'os'
 
 const UPLOAD_DIR = path.resolve(process.env.ROOT_PATH ?? os.tmpdir(), "public/uploads")
 
+function isFile(obj: any): obj is File {
+    return obj instanceof File
+}
+
 export const POST = async (req: NextRequest) => {
     const formData = await req.formData()
     const body = Object.fromEntries(formData)
-    const file = (body.file as Blob) || null
 
-    if (file) {
-        const buffer = Buffer.from(await file.arrayBuffer())
+    if ('file' in body && isFile(body.file)) {
+        const buffer = Buffer.from(await body.file.arrayBuffer())
         if (!fs.existsSync(UPLOAD_DIR)) {
             fs.mkdirSync(UPLOAD_DIR)
         }
 
         fs.writeFileSync(
-            path.resolve(UPLOAD_DIR, (body.file as File).name),
-            buffer
+          path.resolve(UPLOAD_DIR, body.file.name),
+          buffer
         )
-    } else {
         return NextResponse.json({
-            success: false,
+            success: true,
+            name: body.file.name,
         })
     }
 
     return NextResponse.json({
-        success: true,
-        name: (body.file as File).name,
+        success: false,
     })
 }
